+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - Fixed some minor issues
 - Added Russian translation (rfin0)
 - Added option to automatically send commands such as /cp worlds on server join
+- Added more information to config
+- Made command feedback translatable